moves country list to separate component
